Better keyrange management for resharding.
Now the split snapshot manifest contains the keyrange of the
server that took the snapshot, as well as the part that was
snapshot.

blp_recovery now contains the keyrange of the source shard.
That way we can find a server for it.
The player side just computes the overlap before asking the
server.