Handling zookeeper.STATE_CONNECTING as an error.
It turns out that if we go to zookeeper.STATE_CONNECTING state,
and we issue a query, the query fails with 'connection loss'
(that's good), but the session event handler never gets called again,
and the session is stuck in a broken state (that's really bad).
It seems if we don't issue any query before zookeeper.STATE_CONNECTED,
the connection recovers and works as expected (that's good).
We have no way however to distinguish between the two cases.

(note zkocc already does this, and it works much better it seems).