New EventStreamer API for update stream
The new API is different from binlog_parser because
it only transmits the group id. Also, the API now
sends differnt object types (DDL, DML, or position).
The new position is only sent at the end of a transaction.
This ensures that retries will cause the events from the
current transactions to be resent in case of mid-transaction
failures.