queryservice: Change allowQueries
allowQueries is now simplified. It just tries to start the
query service. If it fails, it returns an error. All the
retry logic has been removed. This will allow the tablet manager
to retry based on other conditions that the query service does
not care about.
For vtocc, a failure to start the query service will cause the
program to exit.