binlog_streamer: bug fix for skipping events
The binlog streamer sometimes reads incomplete events due to a race
between the writer and reader. In such cases, it obtains a new position
but fails to read the corresponding event. If this happens, then a
relaunch skips the event because the last position was already recorded.
In this fix, we separate out what we've seen from what we've
transmitted, and always restart mysqlbinlog based on what we've
transmitted, which is always at a transaction boundary.