Add richer counters for GetEndPoints vtgate call.
We want counts split-out by cell/keyspace/shard/dbtype. We also want to
count stats based on what was returned regardless of what was in cache.
In this way, we can reasonably estimate for example, the rate of queries
served errors; this is a better monitoring strategy than reporting
whether there would be an error were a query to be served at the time of
stats poll.