query service: improve schema reload code
Schema reload uses the timestamp of the latest table to figure
out what to load next. This doesn't work well if the
condition is >= because it ends up reloading the
last table all the time.
The new code uses the last timestamp of the server to figure
out what to load next.
Also changed CreateTable to be CreateorUpdate table and changed
usage accordingly.