rowcache: concurrentize delayed spot-checking
If a spot check fails, the rowcache sleeps for 10 seconds
and rechecks the data. This is to accommodate race conditions
that inherently exist between rowcache and database.
Such sleeps increase query execution time. With this CL,
the recheck is launched as a separate goroutine using the
new Launch service of QueryEngine which will ensure that
it won't be closed until such functions have completed.