Remove servenv.exportBinaryVersion.
servenv.exportBinaryVersion md5sums the binary it is in at startup. This
meaningfully impacts startup time for all vitess binaries, since we need
to load several MiB of otherwise unused data from spinning disk.

This is not the appropriate way to manage binary versions. Attaching
this version metadata at build-time with the linker is much better for this.
An example of this is github.com/weingart/vendor, which we should
consider using.