add doc about the life of a query
Life of a query explains the query path from client code to vtgate
to vttablet. It also explains the interactions with topo server, how
the query is handled by vttablet and how the results are returned.

Both regular and streaming queries are included, as well as the
scatter quieries and how it is processed in vtgate.