Wait for mysqlctld to terminate before continuing teardown.
Another race condition exposed by RAM disk:
We were sending SIGTERM to mysqlctld, but not waiting for it.
Then we proceed to delete all the data files. If the deletion is fast
enough, it can remove files that mysqlctld needs to do its teardown, and
mysqld processes are left orphaned.