Speed up TabletExternallyReparented.
This adds an alternate implementation of TabletExternallyReparented
that reduces the client-visible downtime for a reparent by making sure
the serving graph sees the new master ASAP. The rest of topology is
then reconciled as necessary.

The new implementation is only used when the -fast_external_reparent
flag is enabled on vttablet. This flag is currently off by default.