Fix "invalid heap pointer" crash when built with Go 1.4.
The CGO wrapper for ZK was casting arbitrary integers to unsafe.Pointer.
For example, it would call C.something(unsafe.Pointer(1)) in order to
call "void something(void* tag)" with a tag value of (void*)1.

In Go 1.4, if a GC happens to occur while C.something() is executing, it
will find the unsafe.Pointer(1) on the stack and crash because 1 is not
a valid memory address.