Change capitalization of MySQL56 flavor ID.
It should match the convention established with "MariaDB".

I looked into making them case-insensitive, but it would be more tricky
than it's worth because flavor IDs are used as part of the protocol
between servers. So an old server that is case-sensitive would choke on
a message from a new one.