java/vtgate-client: Process new "Flags" MySQL field in result set.
This change breaks the previous behavior: Before, BIGINT values were
always mapped to an UnsignedLong by default, unless they were negative.
With this change, they will be a UnsignedLong or Long depending on the
SQL schema (UNSIGNED flag present or not).

If an old VtGate does not send the Flags field, such values will be
mapped to a Long instead of an UnsignedLong before.

Refactored "FieldType" enum and included it in the "Field" class.