do not check error returned from use database call in ExecuteFetchAsDba
Use database statement is executed if caller specifies dbname. However,
sometimes the given query is trying to create this database and error
will be returned when executing the use database statement. It is okay
to ignore this error because the next ExecuteFetch will run the query and
it will also fail.