Check for unconverted Markdown lists after running Jekyll.
If the writer doesn't put a newline before the list, it might look fine
in GitHub, but get passed through as plain text in Jekyll.

It's hard to detect this case in the raw Markdown without actually
parsing Markdown, because we'd need to distinguish between the first
item of a list, and subsequent items.

However, after running Jekyll, all Markdown lists should have been
converted, so we look for any unconverted lists before publishing.