Check error on Close() and Flush() in Backup.
We are probably doing a bad job of checking these errors in a lot of
places, but it's particularly important to get it right for backups.

We don't worry about Close() on the Reader side.

When using an AllErrorRecorder, we don't have to worry about hiding the
Write error with a Close error. But we do need to make sure not to call
Close as part of saving the arguments of a defer statement.