vtworker: Add Ping command.
A simple vtworker command without any dependencies allows us to easily test the general vtworker framework without having other dependencies working (topology, tablets).

The Ping command accepts a message as input which will be logged with level CONSOLE. In case of an RPC, the complete logging output will be streamed back and therefore the message as well.

The Ping command will not be shown in the "-h" usage output on the command line, but it will be available on the web interface and be triggered from there as well.