Make vtctl print human-readable QueryResults.
The values within each row of a QueryResult were being JSON-encoded as
base64 byte arrays for every type. This JSON encoding was only being
used for tests, until the recent addition of vtctl commands that display
JSON-encoded QueryResults.

Now that we are displaying these to humans, we should display the
results as strings instead of as byte arrays.