worker.py: Created new abstract test base class to fix broken test after moving CopySchemaShard around.
CopySchemaShard cannot be part of the vtworker main test method because
for some test classes the source master is deliberately shut down and
CopySchemaShard will fail.

Instead, I added a new test base class "TestBaseSplitClone" which will
be inherited by the existing "TestBaseSplitCloneResiliency" test class.

Now, CopySchemaShard won't be executed as part of "TestBaseSplitClone"
because it's a step in the resharding process and e.g. the automation
will take care of it.