Fix deadlock in etcdtopo TestWatchEndPoints.
The etcdtopo WatchEndPoints() function was written with the (correct)
assumption that updates can't be missed because the most recent known
index is passed to Watch(), which is supposed to replay any missed
updates.

However, in our fake etcd, we hadn't implemented the replay feature.
Hence, it was possible that the update the test was waiting for would
happen before the watch was installed, leading to a deadlock.