upsert: Make value & last_insert_id as keywords
vttablet cannot support those two functions in upserts.
Marking them as keywords automatically prevents their use.
This way, we don't have to explicitly look for these keywords
inside upserts to prevent them. If/when we decide to support
them, we can add these keywords to the keyword_as_func
rule in the parser.