vtgate test fix
There is a race condition between Balancer and ShardConn
because they both wait for the same amount of time when
a connection is unavailable. If there's only one connection,
this leads to a race condition where you may see a retry
error or a 'no endpoints available' error depending on
which timer wins.
For now, I'm fixing the test such that it avoids that
retry logic. We have to talk about how to make that part
more predictable.