Github front page: Remove coverage button.
The coverage button currently shows 56% coverage for our Go code in a red color.
Unfortunately, this is incorrect because coverage is only tested and
added up for code within the same package. However, our true
coverage is much higher because we have dedicated integration test
packages which cover a lot of other packages.

Therefore, we'll remove the coverage button for now. Once we fixed
calculating the true coverage, we can re-add it.