travis: Custom runner for Go unit tests.
The runner distinguishes between non-flaky and flaky tests.

Non-flaky tests must always pass on the first attempt.

Flaky tests on the other hand are not run in parallel and may be retried up to three times.

I've compiled the blacklist of flaky tests from running the full unit tests hundreds of times.