Deflake vtgatev2_test.
Some of the tests connect to vtgate, then kill the tablet,
and execute a vtgate call that's expected to fail. Out of those,
some of them then go on to try to continue using the vtgate conn.

The problem is that depending on how long it takes to fail, the
first vtgate call might end up hitting a client-side timeout,
which closes the vtgate conn. In the case of gRPC, there is a
set of non-configurable reconnection attempts, which means the
calls can take a while to fail (based on the server-side timeout).

This change increases the client-side timeout for those tests,
to avoid racing with the server-side timeout.