tabletserver: New workflow for SqlQuery
We're changing the workflow for SqlQuery. The old
workflow used to cycle between allowQueries and
disallowQueries.
In the new workflow, we perform an InitDBConnection at the
beginning. Following this, we only call SetServingType. This function
allows you to change the tablet type and whether the query service
should be running or not. This indirectly controls whether the rowcache
invalidator must be started or not. Specifically, we stop the
invalidator if the tablet type is a master.
The old workflow is still supported. We'll look at removing it once
we stop using it. The current implementation supports both workflows.
This means that we have a new state machine that supports all the
above functions.

Some low level details:
Compared to the old state machine, all transient states have
been merged into the following:
StateTransitioning: Any transient state where no queries are allowed.
StateShuttingDown: In this state we allow transactions to complete.
There is also a new state: StateNotConnected. This is the initial state
for vttablet. But we can also be forced into this state if MySQL becomes
unreachable.
For the sake of simplicty and backward compatibility, these states map
to only three state names, SERVING, NOT_SERVING and SHUTTING_DOWN.
Specifically StateNotConnected, StateNotServing, and StateTransitioning
all show up as NOT_SERVING.

CheckMySQL can still be called by rowcache in StateNotServing, and if
MySQL had died during this time, we transition to StateNotConnected.
When we come back up from this state, it's a full initialization
including a brand new schema reload etc. This is equivalent to
allowQueries.

The rowcache invalidator undergoes some tweaks:
1. It's now idempotent. If you start an already running invalidator,
it's a no-op. Same with Stop.
2. If it wasn't previously running, it clears the entire rowcache to
prevent poisoning.

I've also renamed some of the functions:
allowQueries -> StartService
disallowQueries -> StopService
checkMySQL -> CheckMySQL