tabletserver: SqlQuery -> TabletServer
This rename addresses most lint errors in tabletserver.
The last one left is GetSessionId, which will go away once
we migrate to the target-based API.
I've left the bson functions alone to honor any possible legacy
usage. They will also go away as we deprecate those protocols.
As for grpc, I see that the service is already named Query, which
seems fine. I'm not sure if it's worth renaming it. We should talk
about it. So, that part is also unchanged for now.