Add BatchVTGateCursor.executemany.
To comply with PEP0249, remove the BatchVTGateCursor execute and flush
sequence and use executemany instead. The first argument to executemany
is sql if all commands share a common sql (this is supposed to be an
"operation" parameter). The params_list will supply sql if the first
argument is None.

Each params in the executeany params_list generates a result set.
Implement the nextset() to walk through the result_set.

Update tests to use the new BatchVTGateCursor API.