Implement echo tests for goclienttest.
These are ported back from the new Java client tests. Now the Go tests
should be a complete reference implementation of currently available
test cases in vtgateclienttest.

As we add more test cases, we should first add a reference implementation
like this in Go, and then port the tests to other languages.