php: Proof-of-concept protobuf generation.
There is no proto3 compiler for PHP yet, but the wire format of proto3
is backward-compatible with proto2. To avoid maintaining two copies of
our proto files, we use a translator to make the simple syntax changes
necessary to work in PHP's proto2 compiler.

Service stubs are not generated, since I haven't plugged in gRPC yet.