backupstorage: Rename bucket to directory.
The use of the term bucket in the generic backupstorage interface didn't
match the way it's used in GCS or S3. To avoid confusion, we'll call
this the directory, since that's what it actually is in the file-based
case, and in the blobstore case we are emulating the concept of a
directory with object name prefixes.