bootstraph.sh: Fail early if go get fails.
Recently, we saw some Travis failures because not all Go dependencies
were successfully 'go get'd. However, the test run didn't stop early and
instead it was difficult to find out the original reason for the test
failure. This commit will make it easier to find out when the bootstrap
flaked and the test simply has to be restarted.