test/schema: Fix shard_2 setup/teardown.
The test requires that some steps occur before shard_2 comes up,
so shard_2 has a separate setup function. However, the test was trying
to combine teardown of shard_2 with the rest, which was confusing and
resulted in attempting to init mysql twice without tearing it down.

This splits out teardown of shard_2 so it behaves correctly and is
easier to understand.