stats: Remove the stats variable type "States".
Track the state of each Zookeeper connection in the "ZkCachedConn" variable instead.

Remove "ZkMetaConn" variable because "ZkCachedConn" has the same information now.

Update tabletmanager.py end-to-end test.

I'm removing states.go because it caused a deadlock in the worker.py test.

Its removal is no loss because the original intent was to detect flip-flopping with that code. However, that's no longer necessary.

The observed deadlock occurred when a) somebody polls /debug/vars while b) we create a new Zookeeper connection and publish a "ZkCachedConn" variable for it.

When the connection gets created, then the "states" object is created. The same call also calls expvar.Publish() eventually and this is where the deadlock occurs. (It's a deadlock between a mutex in the "expvar" package and a mutex in the Zookeeper connection package.)