test: Unify wait_for_tablet_type_change() helper methods.
There were two almost identical methods in utils.py and in tabletmanager.py.

For the tablet type I'm using strings instead of the proto constants now because it's easier to read and shorter. The proto function which converts the string into an enum value will still check if the type is valid.