vitessdriver: Remove comment that db.Ping() would not be required.
I confused sql.Open() with sql.Driver.Open(). The first one is the general Go interface which still requires db.Ping() to check if a connection works. Driver.Open() is not related to this at all. (It gets called when the Go sql package actually opens the connection e.g. when db.Ping() is called.)