vtworker: Log output of vtworker commands executed via RPC.
One vtworker command may be very long running (on the order of days). If the command is executed via an RPC, the output of the command is currently only stored in memory. There are several situations where this output may get lost halfway:
- if vtworker crashes
- if the consumer crashes and vtworker gets reset

From an operations perspective, we want avoid this and preserve the output instead.

Therefore, the output of each RPC will be logged to the console (and consequently, depending on the flags of the Go log package, to a file as well).

NOTE: The previous code also set a NewMemoryLogger() for the Wrangler. But this wasn't actually necessary.