Remove confusing log messages in serving graph rebuild.
These were logged originally to verify that contention was rare.
If it turned out to be common in practice, we should go back to
pessimistic locking.

Contention on the serving graph has indeed been rare, but when it does
happen, these messages are confusing because they seem to indicate a
problem.