v3: analyze FROM clause
This is the first installment for the new v3 analysis. This change
analyzes just the FROM clause, without the ON clauses, and identifies
possible chunks. For now, chunks are possible only for unsharded
keyspaces. Once we add ON clause analysis, we'll be able to find
chunks for sharded keyspaces also.
The new functions are currently not called from anywhere. There
are some basic unit tests for now, which will also be evolved as
functionality grows.