tabletmanager: Fix problem that reparent of a tablet with active filtered replication failed.
I introduced the issue in f86bb90119c010a5fa28ffcf7540f8d24b513751 where
I changed discovery.WaitForEndPoints() to require a Go context.

In case of a reparent (e.g. moving away from the current master to
slave/spare), the binlog player will be reset/stopped.

As part of that, its context will be cancelled. The same context is used
within Iteration() to wait for the list of healthy endpoints
(WaitForEndPoints()). Due to the cancelled context, WaitForEndPoints() will always return with an error and
therefore Iteration() will
never break out of Loop(). To fix this, I'm checking at the beginning of
Iteration() if the context is still valid.

I've also extended WaitForEndPoints to return a different error when it
actually did not timeout.