v3: add metadata to sqlparser.ColName
When symbols get resolved, those decisions need to be persisted.
Otherwise, the symbol table evolves and introduces new symbols
that may hide old ones (when SELECT is parsed). A re-resolution
after the fact can produce inconsistent/incorrect results.
In other words, we need the original resolution to be remembered
when we decide to wire things up at the end.