java: Remove "notInTransaction" parameter from transaction Execute* methods.
The underlying protobuf field is not used in open-source and therefore we don't want to expose it in our API.

Since the protobuf field will no longer be set, proto3 should always return "false" as zero value for it.

This commit also removes the "asTransaction" flag from the transaction object ExecuteBatch* methods because vtgate ignores this flag anyway when the request is within a transaction.