v3: fix nested subquery merge issue
For more than one level of subquery nesting. The route symbol
table was getting repointed only at the first level. In order
to address this, I'm using a new redirect scheme where a route
can redirect to another route.
In the new scheme we just redirect the current route to the new
one. So, if there are other routes that are redirected to the
current one, they will also get redirected to the new route.