Straw man prototype for V3 refactor.
This started as some notes I was taking to try to understand the
implications of the proposed refactor. I decided to put it up as a straw
man to start pushing the discussion into more concrete terms. It's not
intended to be efficient or even functional. It's just a
whiteboard-level proposal of how we might try to organize the ideas
we've been discussing, typed up instead of written by hand.