v3: extend '*' expr support for postprocessing
If we allowed '*' expressions in select, the postprocessing
col refs can match them a-la-wildcard. There is no harm in
this matching because no column numbers get used int the
process, which is where things could break.
However, we allow '*' expressions only if we know that we
won't need column numbers. So, all is well.