Adding preliminary support for TLS for gRPC.
The server side takes a cert and key, and a ca to verify clients.
The client side takes a cert and key, a ca to verify the server,
and an optional server name in case the server host name does not
match the certificate name.

Renaming secure.py to be more descriptive.