Wiring in immediate callerid for gRPC.
Now extract the gRPC TLS certificate name the client used,
and use that as immediate caller id. When table ACLs are used,
this in effect propagates the certificate name as the client name.

Adding a simple unit test case in secure_transport.py: one vtgate
client cert is allowed to access the test table, another in not.

(also fixing doc for format of table acl config file).