Various fixes after the gRPC secure changes.
vtgateclienttest has to register services in 'OnRun',
instead of right away, as the gRPC server is not created yet.

Various fixes related to socket file. We were wrongly assuming
we needed the gRPC server when a port is specified, forgetting
the unix socket case. Refactoring the unix socket case to be
better aligned with gRPC code.