Simplifying split diff worker.
We used to think we should run one diff per destination shard. Now we
assume we want one diff for each (source, destination) tuple. That way
splits and merges will work the same way, and we never need merge-sort.

In the process, fix the keyrange overlap usage to work for both splits
and merges.