docker: Keep vitess tree from bootstrap.
This is needed so we don't clobber the new vendor directory.

We used to need to clear the tree from bootstrap, because otherwise
files that had been deleted since the last bootstrap image was built
would remain. However, we now only copy a minimal set of files into the
bootstrap image (just enough to run bootstrap.sh), so this is no longer
a concern for ordinary source files.