Enable replication lag check in examples.
Previously, I disabled this check because it was too aggressive.
Any time replication was stopped, it would take the tablet unhealthy.
That meant any problem with the master would take out all tablets
in the whole shard.

In order to get back the replication lag reporting in realtime stats,
I've removed the check that replication is running and re-enabled the
replication lack health reporter.

The replication lag is returned as 0 when not replicating, but this
should be a better compromise than having no lag info at all,
in the absence of a more resilient heartbeat-based measurement.