InitTablet: Don't need shard lock to add cell.
I found a race with ZK when many tablets try to init at the same time.
In ZK, creating the shard consists of creating multiple paths
(the shard record itself, /action, etc.). If the second tablet to init
sees that the shard record exists, but tries to lock the shard before
the /action path exists, it will fail.

As a workaround, I've changed the shard lock to use UpdateShardFields
because the lock is not required according to our revised locking
guidelines. We only need to ensure the consistency of the shard record
with itself, and UpdateShardFields can do that. This gets around the ZK
race because it doesn't depend on any of the secondary paths existing.