Fixing health check.
The aggregator was not preserving the special error case, now it does,
and has unit tests to prove it.

The integration tests were not fully resetting replication (using 'reset
slave' and not 'reset slave all'). The difference is 'reset slave'
clears the values from 'show slave status', whereas 'reset slave all'
makes the status show nothing (as is the startup case).