Making keyspace_id check optional in vttablet.
It's only enabled now if the Keyspace object has a ShardingColumnName
set. We do not require the ShardingColumnName field any more.

For v3, do not send a hardcoded keyspace_id any more. That was just a
hack.

The worker tests do not need to set ShardingColumnName any more.
v3 test doesn't need to anymore either.