tabletmanager: Remember if we've been told to stop replicating.
We remember the setting on the file system, because we want it to work
regardless of whether mysqld is running, and because if that tablet dir
is lost, the particular position at which we're stopped is also lost.
Therefore, the setting to stay stopped should not persist across loss of
the tablet dir (e.g. it should not be stored in topology).