healthcheck: Try to heal replication automatically.
This allows tablets (consisting of vttablet+mysqld) to be restarted
without having to manually re-point slaves at the master and tell them
to start replicating again. This is important for rollout flows, where
each slave is restarted in turn.

We can't just check at vttablet startup, because then we'd miss the case
when only mysqld is restarted. Instead, we check at each healthcheck
interval so replication becomes self-healing. The replication lag check
had to be moved from mysqlctl to tabletmanager and integrated with this
healing, to make sure the healing attempt is synchronized with the slave
status check.

This is also needed for Orchestrator integration, because we are responsible
for giving the slave a master in the first place so Orchestrator knows which
replication tree the slave belongs to.