Enable relay_log_recovery for MySQL 5.6+.
This causes mysqld to ignore relay logs on disk at startup,
and instead start replicating from the master wherever the SQL thread
left off.

This is generally recommended for safer crash recovery.
We also do it to avoid problems caused by spurious Previous-GTIDs lists
in relay logs that occur when we restart mysqld in integration tests.