discovery: Guarantee that HealthCheck stops calling the listener after HealthCheck.Close() returned.
Without this change, it's difficult to synchronize the shutdown of a listener implementation and HealthCheck.

For example, a listener implementation could send to a channel for every update. But the shutdown of the listener implementation will close that channel and future sends to it will panic. Therefore, a HealthCheck.Close() must guarantee the user that the listener won't be called anymore when it returned. After that, the listener can shut down itself.