Refactoring vtgatev2_test a bit.
Not restarting processes when we don't have to.
Closing vtgate connections when not used.
Not waiting for SERVING state (when the start_vttablet does it already).
Killing vtgate explicitely at the end to avoid spam.

Execution time went from 210s to 145s on my WS.

Also swallowing execptions on the rollback() in close() for GRPC vtgate
client. We want to close that connection after all.