Randomize MySQL server-id when initializing tablet from scratch.
When creating a tablet from an empty data dir (mysqlctl init),
we should pick a random MySQL server-id, rather than using the tablet
UID as the server-id. This better matches the expectations of MySQL that
a server with the same server-id ought to be the same incarnation of the
data dir. Violating that expectation causes bad interactions with
features like ignoring binlog events from the same server-id (when
replicate-same-server-id=FALSE, which is the default).

If a tablet is just being restarted in-place with its data dir retained
(mysqlctl shutdown/start), then it will re-use the same server-id that
was previously generated.