tabletserver: enforce AUTOCOMMIT mode
VTTablet will not work correctly if AUTOCOMMIT was OFF. This is
because it uses connection pools. If AUTOCOMMIT is OFF, MySQL
will start a transaction at the beginning, which will create a
snapshot of the database that will never change for the rest
of the life's connection. If AUTOCOMMIT was ON, then there's
an implicit commit after every read, which will reset the snapshot
view.