Removing duplicate parameters from tabletconn.
The Dialer was taking a Tablet, keyspace, shard and tablettype. These
last three are all in the tablet too, so just using these is easier.

In the fakehealthcheck, we were initializing a sandbox conn with no
tablet. Then creating the fake tablet, and never setting the sandbox
conn tablet. Reversed the order of construction, so now we create the
tablet first, and the sandbox conn to it.