worker: Call RefreshState instead of ReloadSchema to start binlog players.
ReloadSchema wasn't really the right thing to do here. That should be
done when the schema is actually copied. It was working by coincidence
because ReloadSchema was under RPCWrapLockAction, which triggers
RefreshState. That broke when I changed ReloadSchema to RPCWrap, so I
need to fix worker clone to use RefreshState instead.